Remove decrypt.go
Most data is now unused. Copied important pieces to new files
